Rewrite the glibc build for 2.16, removing packages
This build now looks at the rbraun/hurd_with_pthreads branch.  This is not expected to be a permanent branch, and should be integrated into master soon, but the timing is unknown, and the branch seems usable for testing purposes.  As a side effect, all patches were removed, but due to a slight change pthread.h needs to be included from libpthread.git, alleviating a circular dependency with the hurd package.